ch 3
almost done editing ch 3 (for formatting).  I need to absorb the auxuliary results from the appendix and check the proofs carefully once again. I like the new format where the proofs are presented right after the statments and the various support recovery criteria are presented in separate sections rather than sub-sections.  Thus, they are visible in the table of contents.