simplified a bit
It turns out that we don't need a special condition on the slowly varying function \ell in the case \beta=1.  I just added an argument in the proof of Theorem 4.1 based on Potter bounds.

This simplified the exposition!

Please check, I will carefully check one more time tomorrow.

BTW, I am sure you know, but a safe way to make sure we have no broken references is to examine the log and go page by page through the missing labels.  This is also a way to securely replace all the references to an outdated  equation -- just eliminate its label and go through the errors...